+++ sandbox/modules/identity/src/main/java/org/opensearch/identity/rest/action/AddPermissionAction.java
@@ -11,14 +11,20 @@ package org.opensearch.identity.rest.action;
 import org.opensearch.action.ActionType;
 import org.opensearch.identity.rest.response.AddPermissionResponse;
 
+/**
+ * This class defines the AddPermissionAction ActionType which corresponds to an action that grant a subject permissions while using Identity.
+ */
 public class AddPermissionAction extends ActionType<AddPermissionResponse> {
 
     public static final AddPermissionAction INSTANCE = new AddPermissionAction();
 
     // TODO : revisit this action type
-    public static final String TYPE = "cluster:permission";
+
+
+     // The action name
+    public static final String NAME = "cluster:permission/add";
 
     AddPermissionAction() {
-        super(TYPE, AddPermissionResponse::new);
+        super(NAME, AddPermissionResponse::new);
     }
 }