+++ sandbox/modules/identity/src/main/java/org/opensearch/identity/rest/action/TransportDeletePermissionAction.java
@@ -23,7 +23,7 @@ public class TransportDeletePermissionAction extends HandledTransportAction<Dele
     private final UserService userService;
 
     public TransportDeletePermissionAction(TransportService transportService, ActionFilters actionFilters, UserService userService) {
-        super(AddPermissionAction.TYPE, transportService, actionFilters, DeletePermissionRequest::new);
+        super(AddPermissionAction.NAME, transportService, actionFilters, DeletePermissionRequest::new);
         this.userService = userService;
     }